Fixed loading of test resources
